requirements.txt
tensorflow==2.9.1
tensorflow_hub=0.12.0
streamlit==1.10.0
Pillow==9.1.1
numpy==1.2.4